Hotfix/rollback/v2.1.0 (#230)
* Data update 2025-02-23-21-40-37 [skip ci]

  - Synced data files from main

* Release/v1.6.0 single model evaluations (#160)

* Bring Upstream Main up to Speed & Various Updates (#152)

* Updated data-fetching pipeline

* merged forecast chart margin update + cap prediction line

* fixes and updates

* Default selectedWeek to latest valid prediction-available date

* SettingsStateMap changed to page-agnostic + Change log update + Logo

* fix typo in data retrieval script that broke pipeline (#153)

* fixed pipeline issue (#154)

* Data update 2025-02-23-21-40-37

  - Updated data files
  - Updated submodules reference

* fix to import paths (#155)

* normalized module import path casing to avoid nextjs build issue (#156)

* update to folder names

* added cache invalidation for debugging (#157)

* next.js should fix their case enforcing rules smh

---------


* Data update 2025-02-27-06-38-43 [skip ci]

  - Synced data files from main

* Data update 2025-02-27-15-35-17 [skip ci]

  - Synced data files from main

* Data update 2025-02-27-18-38-27 [skip ci]

  - Synced data files from main

* Revert "Optimization/data processing/pre calculation for evals season overvie… (#223)"

This reverts commit 01cf67ab780d8564c4a358457ba3f7ec8b1bafce, reversing
changes made to 2245aaa884fc6415cf88fa738a515ca47ad80623.

---------
