+++ src/app/evaluations/evaluations-components/SeasonOverview/MapSelectorPanel.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useAppDispatch, useAppSelector } from "@/store/hooks";
-import { modelNames, modelColorMap } from "@/types/common";
+import { modelNames, modelColorMap } from "@/interfaces/epistorm-constants";
 import { setMapSelectedModel, setMapSelectedScoringOption, setUseLogColorScale } from "@/store/evaluations-season-overview-settings-slice";
 
 interface MapSelectorPanelProps {
