+++ src/app/evaluations/evaluations-components/SeasonOverview/SeasonOverviewPIChart.tsx
@@ -6,9 +6,8 @@ import * as d3 from "d3";
 import { addWeeks } from "date-fns";
 
 import { useAppSelector } from "@/store/hooks";
-import { useResponsiveSVG } from "@/utils/responsiveSVG";
-import { modelColorMap, modelNames } from "@/types/common";
-import { selectSeasonOverviewData, selectShouldUseJsonData } from "@/store/selector/evaluationSelectors";
+import { useResponsiveSVG } from "@/interfaces/responsiveSVG";
+import { modelColorMap, modelNames } from "@/interfaces/epistorm-constants";
 
 // Interface for processed data structure
 interface ProcessedCoverageData {
@@ -23,76 +22,31 @@ const SeasonOverviewPIChart: React.FC = () => {
   const { containerRef, dimensions, isResizing } = useResponsiveSVG();
   const chartRef = useRef<SVGSVGElement>(null);
 
-  // Get data from selectors
-  const shouldUseJsonData = useAppSelector(selectShouldUseJsonData);
-  const seasonOverviewData = useAppSelector(selectSeasonOverviewData);
-
   // Get data from Redux store
   const detailedCoverageData = useAppSelector((state) => state.evaluationsSingleModelScoreData.detailedCoverage);
   const { evaluationSeasonOverviewHorizon, selectedAggregationPeriod, aggregationPeriods, evaluationSeasonOverviewSelectedModels } =
     useAppSelector((state) => state.evaluationsSeasonOverviewSettings);
 
-  // Process the detailed coverage data using JSON when available, otherwise CSV fallback
+  // Process the detailed coverage data based on selected criteria
   const processedData = useMemo(() => {
-    if (shouldUseJsonData && seasonOverviewData) {
-      // Use JSON data structure
-      const coverageData = seasonOverviewData.coverageData;
-      const results: ProcessedCoverageData[] = [];
-      
-      // Coverage levels mapping for JSON data
-      const coverageLevels = [10, 20, 30, 40, 50, 60, 70, 80, 90, 95, 98];
-      
-      // Process each selected model
-      for (const modelName of modelNames.filter(m => seasonOverviewData.selectedModels.includes(m))) {
-        const modelCoverageData = coverageData[modelName];
-        if (!modelCoverageData) continue;
-        
-        const coveragePoints: { covLevel: number; coverageValue: number }[] = [];
-        
-        // Calculate coverage for each level by aggregating across horizons
-        coverageLevels.forEach(level => {
-          let totalSum = 0;
-          let totalCount = 0;
-          
-          seasonOverviewData.horizons.forEach(horizon => {
-            const horizonData = modelCoverageData[horizon];
-            if (horizonData && horizonData[level]) {
-              totalSum += horizonData[level].sum;
-              totalCount += horizonData[level].count;
-            }
-          });
-          
-          if (totalCount > 0) {
-            coveragePoints.push({
-              covLevel: level,
-              coverageValue: totalSum / totalCount,
-            });
-          }
-        });
-        
-        if (coveragePoints.length > 0) {
-          results.push({
-            modelName,
-            coveragePoints,
-          });
-        }
-      }
-      
-      console.debug("Using JSON data for PI chart:", results);
-      return results;
-    }
-
-    // Fallback to original CSV processing logic
     if (!detailedCoverageData || evaluationSeasonOverviewHorizon.length === 0 || !selectedAggregationPeriod) {
       return [];
     }
 
+    // Find selected time period
     const selectedPeriod = aggregationPeriods.find((p) => p.id === selectedAggregationPeriod);
-    if (!selectedPeriod) return [];
 
+    if (!selectedPeriod) {
+      return [];
+    }
+
+    // Pre-create set for faster horizon lookups
     const horizonSet = new Set(evaluationSeasonOverviewHorizon);
+
+    // Process data for each model
     const results: ProcessedCoverageData[] = [];
 
+    // Map from coverage field names to confidence levels
     const coverageMapping = [
       { field: "coverage10", level: 10 },
       { field: "coverage20", level: 20 },
@@ -107,27 +61,37 @@ const SeasonOverviewPIChart: React.FC = () => {
       { field: "coverage98", level: 98 },
     ];
 
+    // Process each model's data
     for (const model of detailedCoverageData) {
+      // Skip if no coverageData
       if (!evaluationSeasonOverviewSelectedModels.includes(model.modelName) || !model.coverageData || model.coverageData.length === 0)
         continue;
 
+      // Initialize arrays to store sums and counts for each confidence level
       const coverageSums = coverageMapping.map(() => 0);
       const coverageCounts = coverageMapping.map(() => 0);
 
+      // Filter and aggregate coverage data
       for (const entry of model.coverageData) {
+        // Skip if horizon doesn't match
         if (!horizonSet.has(entry.horizon)) continue;
 
+        // Calculate target date for filtering by time period
         const referenceDate = entry.referenceDate;
         const targetDate = addWeeks(referenceDate, entry.horizon);
 
+        // Skip if outside selected time period
         if (referenceDate < selectedPeriod.startDate || targetDate > selectedPeriod.endDate) continue;
 
+        // Now this entry passes all filters, add its values to the sums
         coverageMapping.forEach((mapping, index) => {
+          // Add value to the sum for this confidence level
           coverageSums[index] += entry[mapping.field as keyof typeof entry] as number;
           coverageCounts[index]++;
         });
       }
 
+      // If we have data for this model, calculate averages and add to results
       if (coverageCounts.some((count) => count > 0)) {
         const coveragePoints = coverageMapping.map((mapping, index) => ({
           covLevel: mapping.level,
@@ -141,17 +105,8 @@ const SeasonOverviewPIChart: React.FC = () => {
       }
     }
 
-    console.debug("Using CSV fallback data for PI chart:", results);
     return results;
-  }, [
-    shouldUseJsonData,
-    seasonOverviewData,
-    detailedCoverageData,
-    evaluationSeasonOverviewHorizon,
-    selectedAggregationPeriod,
-    aggregationPeriods,
-    evaluationSeasonOverviewSelectedModels,
-  ]);
+  }, [detailedCoverageData, evaluationSeasonOverviewHorizon, selectedAggregationPeriod, aggregationPeriods, evaluationSeasonOverviewSelectedModels]);
 
   /* UseEffect Hook for rendering the chart */
   useEffect(() => {
