+++ src/app/forecasts/forecasts-components/NowcastStateThermo.tsx
@@ -3,7 +3,7 @@ import * as d3 from "d3";
 import * as topojson from "topojson-client";
 import { useAppSelector } from "@/store/hooks";
 import { format, subDays } from "date-fns";
-import { isUTCDateEqual } from "@/utils/date";
+import { isUTCDateEqual } from "@/interfaces/forecast-interfaces";
 
 const shapeFile = "/states-10m.json";
