+++ src/app/interfaces/epistorm-constants.ts
@@ -1,0 +1,15 @@
+// Color Mapping for each model
+
+export const modelColorMap: Record<string, string> = {
+    "MOBS-GLEAM_FLUH": "#9CEB94",
+    "MIGHTE-Nsemble": "#95F4C9",
+    "MIGHTE-Joint": "#3FC49E",
+    "NU_UCSD-GLEAM_AI_FLUH": "#45CDED",
+    "CEPH-Rtrend_fluH": "#0292D1",
+    "NEU_ISI-FluBcast": "#7BB1FF",
+    "NEU_ISI-AdaptiveEnsemble": "#5F5FD6",
+    "FluSight-ensemble": "#D36F54",
+//     Add more here in the future...
+};
+
+export const modelNames: string[] = ["MOBS-GLEAM_FLUH", "MIGHTE-Nsemble", "MIGHTE-Joint", "NU_UCSD-GLEAM_AI_FLUH", "CEPH-Rtrend_fluH", "NEU_ISI-FluBcast", "NEU_ISI-AdaptiveEnsemble", "FluSight-ensemble"];
