+++ src/app/store/data-slices/locationSlice.ts
@@ -1,6 +1,6 @@
 // src/app/store/locationSlice.ts
-import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { LocationData } from '@/types/domains/forecasting';
+import {createSlice, PayloadAction} from '@reduxjs/toolkit';
+import {LocationData} from '@/interfaces/forecast-interfaces';
 
 interface LocationState {
     data: LocationData[];
