+++ src/app/store/data-slices/nowcastTrendsSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { NowcastTrendByModel, NowcastTrendsCollection } from '@/types/domains/nowcast';
+import {NowcastTrendByModel, NowcastTrendsCollection} from '@/interfaces/forecast-interfaces';
 
 
 const initialState: NowcastTrendsCollection = {
