+++ src/app/store/evaluations-single-model-settings-slice.ts
@@ -1,6 +1,6 @@
 // src/app/store/forecastSettingsSlice.ts
 import {createSlice, PayloadAction} from '@reduxjs/toolkit';
-import { SeasonOption } from '@/types/domains/forecasting';
+import {SeasonOption} from '@/interfaces/forecast-interfaces';
 import {parseISO} from "date-fns";
 
 interface EvaluationsSettingsState {
