+++ src/app/store/index.ts
@@ -1,17 +1,16 @@
 import { configureStore } from "@reduxjs/toolkit";
 
+// Import Data Reducers
 import groundTruthReducer from "./data-slices/groundTruthSlice";
 import predictionsReducer from "./data-slices/predictionsSlice";
 import locationReducer from "./data-slices/locationSlice";
 import nowcastTrendsReducer from "./data-slices/nowcastTrendsSlice";
 import stateThresholdsReducer from "./data-slices/stateThresholdsSlice";
 import historicalGroundTruthReducer from "./data-slices/historicalGroundTruthSlice";
+
 import evaluationsSingleModelScoreDataReducer from "./data-slices/evaluationsScoreDataSlice";
 
-// Import All-in-one new Evaluation Data Slice
-import evaluationDataReducer from "./data-slices/evaluationDataSlice";
-
-// Import settings reducers
+// Import Settings Reducers
 import forecastSettingsReducer from "./forecast-settings-slice";
 import evaluationsSeasonOverviewSettingsReducer from "./evaluations-season-overview-settings-slice";
 import evaluationsSingleModelSettingsReducer from "./evaluations-single-model-settings-slice";
@@ -21,24 +20,17 @@ const store = configureStore({
     groundTruth: groundTruthReducer,
     predictions: predictionsReducer,
     location: locationReducer,
+    forecastSettings: forecastSettingsReducer,
     nowcastTrends: nowcastTrendsReducer,
     stateThresholds: stateThresholdsReducer,
     historicalGroundTruth: historicalGroundTruthReducer,
-
-    forecastSettings: forecastSettingsReducer,
-
-    evaluationData: evaluationDataReducer,
-
-    evaluationsSingleModelScoreData: evaluationsSingleModelScoreDataReducer,
-
     evaluationsSeasonOverviewSettings: evaluationsSeasonOverviewSettingsReducer,
     evaluationsSingleModelSettings: evaluationsSingleModelSettingsReducer,
+    evaluationsSingleModelScoreData: evaluationsSingleModelScoreDataReducer,
   },
 });
 
 export type RootState = ReturnType<typeof store.getState>;
 export type AppDispatch = typeof store.dispatch;
-
-export * from './selector/evaluationSelectors';
 
 export default store;
