--- src/app/types/app.ts
@@ -1,13 +1,0 @@
-export interface LoadingStates {
-  groundTruth: boolean;
-  predictions: boolean;
-  locations: boolean;
-  nowcastTrends: boolean;
-  thresholds: boolean;
-  historicalGroundTruth: boolean;
-  seasonOptions: boolean;
-  evaluationScores: boolean;
-  evaluationDetailedCoverage: boolean;
-}
-
-
