--- src/app/utils/date.ts
@@ -1,9 +1,0 @@
-export const isUTCDateEqual = (a: Date, b: Date) => {
-  return (
-    a.getUTCFullYear() === b.getUTCFullYear() &&
-    a.getUTCMonth() === b.getUTCMonth() &&
-    a.getUTCDate() === b.getUTCDate()
-  );
-};
-
-
